day6: Refactor code creating domain and finishing unit test for pods client
